Consolidate under getViewByURL
